requirements.txt
Django
psycopg2
requests
south
django-localflavor
django-geojson
django-haystack
djangorestframework
git+https://github.com/makinacorpus/django-leaflet.git#egg=django-leaflet
mock
numpy
pyelasticsearch
jsonschema
